Switching requirements.py to updated file
